Add ConvMBChrsToAmiga() and ConvMBChrsFromAmiga() to nikom.library.
API wise these have the same function as ConvChrsToAmiga() and
ConvChrsFromAmiga(), but support multi byte character sets where
ISO-8859-1 characters are encoded as upto 2 bytes. In order to do this
in a clean way the new functions take separate source and destination
buffers as arguments.

Implementation wise support for the UTF-8 character set is added.