Initial UTF-8 support for NiKom (#88)
* Add CHRS_UTF8 define.

* Add ConvMBChrsToAmiga() and ConvMBChrsFromAmiga() to nikom.library.

API wise these have the same function as ConvChrsToAmiga() and
ConvChrsFromAmiga(), but support multi byte character sets where
ISO-8859-1 characters are encoded as upto 2 bytes. In order to do this
in a clean way the new functions take separate source and destination
buffers as arguments.

Implementation wise support for the UTF-8 character set is added.

* Use the new ConvMBChrs* functions in SerialIO to support UTF-8.

* Allow user character set to be changed to UTF-8.

* Update MSG_CHRS_BAD for new choice range 1-5.

Also move the actual range from translated string to the code, so
updating it is not forgotten in the future.

* Mention UTF-8 support in ReleaseNotes.

* Remove never used PS_NOCONV flag.

* ConvMBChrs*: Do not check charset for every character.

Check character set once and call appropriate function handling the
entire string. Avoids both checking character set and dereferencing
NiKomBase for each character.