Add Fido support for UTF-8, CP850, CP866 and modern CHRS names (#92)
* ReadFidoText: Use correct length for "MAC 2" charset match.

* ReadFidoText: Do not continue comparing CHRS string on match.

* ReadFidoText: Read and write modern FTS-5003 CHRS kludge names.

IBMPC is a deprecated name, initially meant to mean CP437, but later
used for any CP* charset.
MAC is a deprecated name for CP10000.

* Formatting cleanup.

* Rename convCP850ToAmiga to convNoKludgeToAmiga.

It's not really CP850 we are handling, just a small subset.
This makes the ConvMBChrs* functions behave just like the old
ConvChrs* functions when passed 0 as the chrs.
CHRS_CP850 is left reserved for proper handling of CP850.

* NiKomBase.h: Indent according to current coding style.

* CharTabs.h: Re-indent to keep lines less than 80 characters.

* CharTabs.h: Make arrays static const.

* Add simple python utilty to create the basis of charset tables.

Generates mapping tables for any one-byte-per-character character set
supported by python. Saves doing the trivial but boring parts when
creating new mapping tables.

* Add full support for CP850.

CP850 is still used in FidoNet, and this allows us to fully handle
properly marked CP850 texts, instead of using the convnokludge hack.

* Change CHRS_CP850 to 3.

No reason to have the CHRS_ constants as a bitmask. The other
constants are kept as is, as there is existing config using them.

* Add support for CP866.

This is one of the character sets used in FidoNet.

* Make convUTF8ToAmiga() non-static.

* Initial support for reading UTF-8 Fido texts.

* Add dstlen argument to ConvMBChrsFromAmiga.

When we may need to truncate output we must ensure that we do not make
the truncation in the middle of an UTF-8 character, so make the
conversion function aware of the maximum destination size so it can
make sure to only write complete characters.

* Use ConvMBChrs* functions in ReadFidoText.c

Allows reading and writing multi byte character sets.

* Make local functions static.

* Convert CHRS_ defines to enum nikom_chrs.

Using an enum in switch statements allows gcc to warn on unhandled
enum values.
Also create explicit constant CHRS_UNKNOWN for the 0 value.

* Add emacs backup files to .gitignore