Fixing so that you can change the subject of Fido texts/mail in the editor.
This involved some cleaning up that introduced an abstraction struct that
contain pointers to data that can be changed inside the editor. The editor
no longer needs to be aware of what kind of text that is being written and
doesn't need direct access to the data structures of the texts being written.

Fixes #128